use of int32_t for the dimension + adaptation of the general function in distance.h

+++ headers/kMeans.h
@@ -6,7 +6,7 @@
 typedef struct {
     cluster_t *clusters;
     point_t **points;
-    int64_t dimension;
+    u_int32_t dimension;
     int64_t size;
 } k_means_t;
